+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 1.22.7 (2024-tba-tba)
+
+## Updates/Fixes
+- fixed ledger trade asset bug [#427](https://github.com/asgardex/asgardex-desktop/pull/428)
+- fixed cosmos url [#413] (https://github.com/asgardex/asgardex-desktop/pull/418)
+- fixed btc/ uxto max send [#434] (https://github.com/asgardex/asgardex-desktop/pull/437)
+- fixed deposit page bug [#426] (https://github.com/asgardex/asgardex-desktop/pull/430)
+- udpate error handling in savers [#424] (https://github.com/asgardex/asgardex-desktop/pull/441)
+
 # 1.22.6 (2024-10-6)
 
 ## Updates/Fixes